--- a/.gitignore
+++ b/.gitignore
@@ -59,3 +59,6 @@ binaries
 
 # misc
 .DS_Store
+
+# vscode
+.vscode