<!--  Thanks for sending a pull request! -->

**What this PR does / why we need it**:

There was two problem in func deleteNodes.
1. repeat splay for each node to be deleted resulting in numerous rotations.
2. not actually delete node, zero-weight node left and affect.

improve the deletion logic :
to skip splay for a node in between, and to isolate in a subtree then cut off the subtree.

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Fixes #

**Special notes for your reviewer**:

delete-all bench was too small so I increased the size by 10 times.

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".
-->
```release-note

```

**Additional documentation**:

<!--
This section can be blank if this pull request does not require a release note.

Please use the following format for linking documentation:
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```

**Checklist**:
- [x] Added relevant tests or not required
- [x] Didn't break anything