@hackerwins 2022-06-10
This PR has the following performance improvements compared to main branch(https://github.com/yorkie-team/yorkie/commit/936525f3aecf0c9e5d50a742ceeb0cb9bcde3f28):

Before:

  - Commit: https://github.com/yorkie-team/yorkie/commit/936525f3aecf0c9e5d50a742ceeb0cb9bcde3f28
  - Action: https://github.com/yorkie-team/yorkie/runs/6827624090

```
text_delete_all_10000-2    	      19	  56862292 ns/op	16464676 B/op	  580718 allocs/op
BenchmarkTextEditing-2                 1	30972049507 ns/op	8396292624 B/op	16986104 allocs/op
```

After:

  - Commit: https://github.com/yorkie-team/yorkie/pull/341/commits/8113662d90c3f00892415e0c1d33f4d5a2bf0ffe
  - Action: https://github.com/yorkie-team/yorkie/runs/6828407409

```
text_delete_all_10000-2    	      70	  16737826 ns/op	 5963835 B/op	   71108 allocs/op
BenchmarkTextEditing-2        	       1	24288249925 ns/op	8396771312 B/op	17239469 allocs/op
```